REFACTOR: move files around to conform to jamstik
