+++ src/openBCIConstants.js
